moved action buttons (make proposal, retract....) from the connection-context-dropdown to the chat-textfield
